backend/requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
numpy==1.26.0
packaging==23.2
pandas==1.5.0
pinecone-client==2.2.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==2.3.3
WTForms==3.0.1
flask-jwt-extended==4.5.2